Cleanup parsers
