+++ Source/FunicularSwitch.Generators/ExtendMonad/Parser.cs
@@ -1,5 +1,6 @@
 using FunicularSwitch.Generators.Common;
 using FunicularSwitch.Generators.Generation;
+using FunicularSwitch.Generators.Parsing;
 using FunicularSwitch.Generators.Transformer;
 using Microsoft.CodeAnalysis;
 
@@ -7,23 +8,19 @@ namespace FunicularSwitch.Generators.ExtendMonad;
 
 internal class Parser
 {
-    public static GenerationResult<ExtendMonadInfo> GetExtendedMonadSchema(INamedTypeSymbol targetSymbol, ExtendMonadAttribute extendMonadAttribute, CancellationToken cancellationToken)
-    {
-        var accessModifier = Transformer.Parser.DetermineAccessModifier(targetSymbol);
-        return
-            from monadInfo in Transformer.Parser.ResolveMonadDataFromMonadType(targetSymbol, cancellationToken)
-            let staticMonadGenerationInfo = new StaticMonadGenerationInfo(
-                targetSymbol.Name,
-                accessModifier,
-                [],
-                MonadMethods.CreateExtendMonadMethods(
-                    monadInfo.GenericTypeName,
-                    monadInfo
-                )
+    public static GenerationResult<ExtendMonadInfo> GetExtendedMonadSchema(INamedTypeSymbol targetSymbol, ExtendMonadAttribute extendMonadAttribute, CancellationToken cancellationToken) =>
+        from monadInfo in MonadParser.ResolveMonadDataFromMonadType(targetSymbol, cancellationToken)
+        let staticMonadGenerationInfo = new StaticMonadGenerationInfo(
+            targetSymbol.Name,
+            targetSymbol.GetActualAccessibility(),
+            [],
+            MonadMethods.CreateExtendMonadMethods(
+                monadInfo.GenericTypeName,
+                monadInfo
             )
-            select new ExtendMonadInfo(
-                targetSymbol.FullTypeNameWithNamespace(),
-                targetSymbol.GetFullNamespace(),
-                staticMonadGenerationInfo);
-    }
+        )
+        select new ExtendMonadInfo(
+            targetSymbol.FullTypeNameWithNamespace(),
+            targetSymbol.GetFullNamespace(),
+            staticMonadGenerationInfo);
 }
